# Swagger/api_mod.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API",
    "description": "API for Portal System in AEM",
    "termsOfService": "None"
  },
  "host": "otvcrm01:84",
  "basePath": "/",
  "schemes": [ "http", "https" ],
  "consumes": [ "application/json" ],
  "produces": [ "application/json" ],
  "paths": {
...
    "/api/{lang}/FundingOpportunities/GetByNameAsync/{title}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetByNameAsyncByTitleGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      },
      "post": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "fundingOpportunity",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetAllFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetAllFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetActiveFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetActiveFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetAwaitingPublishedFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetAwaitingPublishedFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetDraftFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetDraftFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetInactiveFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetInactiveFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetArchivedFundingOpportunities": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetArchivedFundingOpportunitiesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetAllFundingOpportunitiesByProgram/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetAllFundingOpportunitiesByProgramByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetActiveFundingOpportunitiesByProgram/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetActiveFundingOpportunitiesByProgramByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetAwaitingPublishedFundingOpportunitiesByProgram/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetAwaitingPublishedFundingOpportunitiesByProgramByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetDraftFundingOpportunitiesByProgram/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetDraftFundingOpportunitiesByProgramByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetInactiveFundingOpportunitiesByProgram/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetInactiveFundingOpportunitiesByProgramByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetArchivedFundingOpportunitiesByProgra/{programId}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetArchivedFundingOpportunitiesByPrograByProgramIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/GetActiveFundingOpportunitiesWithRelationshipsCollapsed": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesGetActiveFundingOpportunitiesWithRelationshipsCollapsedGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunity" }
            }
          }
        }
      }
    },
    "/api/{lang}/FundingOpportunities/{id}": {
      "get": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          }
        }
      },
      "put": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "fundingOpportunity",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "FundingOpportunities" ],
        "operationId": "ApiByLangFundingOpportunitiesByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunity" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FundingOpportunityInternalUser": {
      "type": "object",
      "properties": {
        "fundingOpportunityInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "hasFullAccess": { "type": "boolean" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "FundingOpportunityEligibilityCriteria": {
      "type": "object",
      "properties": {
        "fundingOpportunityEligibilityCriteriaId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "eligibilityCriteriaId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "EligibleClientType": {
      "type": "object",
      "properties": {
        "eligibleClientTypeId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "eligibleClientTypeBilingualTitle": { "type": "string" },
        "eligibleClientTypeStaticId": {
          "format": "int32",
          "type": "integer"
        },
        "titleE": { "type": "string" },
        "titleF": { "type": "string" },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "lang": { "type": "string" },
        "description": { "type": "string" },
        "title": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "EligibleCostCategory": {
      "type": "object",
      "properties": {
        "eligibleCostCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "costCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "titleE": { "type": "string" },
        "titleF": { "type": "string" },
        "tooltipE": { "type": "string" },
        "tooltipF": { "type": "string" },
        "lang": { "type": "string" },
        "title": { "type": "string" },
        "toolTip": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "FundingOpportunity": {
      "type": "object",
      "properties": {
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingProgramId": {
          "format": "uuid",
          "type": "string"
        },
        "titleE": { "type": "string" },
        "titleF": { "type": "string" },
        "lang": { "type": "string" },
        "title": { "type": "string" },
        "description": { "type": "string" },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "objectives": { "type": "string" },
        "expectedResults": { "type": "string" },
        "eligibilityCriterias": { "type": "string" },
        "additionalInformationE": { "type": "string" },
        "additionalInformationF": { "type": "string" },
        "activationStartDate": {
          "format": "date-time",
          "type": "string"
        },
        "activationEndDate": {
          "format": "date-time",
          "type": "string"
        },
        "onHold": { "type": "boolean" },
        "status": {
          "format": "int32",
          "enum": [ 0, 1, 2, 3 ],
          "type": "integer"
        },
        "statusDesc": { "type": "string" },
        "fundingOpportunityExpectedResults": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityExpectedResult" }
        },
        "fundingOpportunityObjectives": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityObjective" }
        },
        "fundingOpportunityEligibilityCriterias": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityEligibilityCriteria" }
        },
        "eligibleClientTypes": {
          "type": "array",
          "items": { "$ref": "#/definitions/EligibleClientType" }
        },
        "eligibleCostCategories": {
          "type": "array",
          "items": { "$ref": "#/definitions/EligibleCostCategory" }
        },
        "fundingOpportunityConsiderations": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityConsideration" }
        },
        "fundingOpportunityFrequentlyAskedQuestions": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityFrequentlyAskedQuestion" }
        },
        "fundingOpportunityResources": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityResource" }
        },
        "fundingOpportunityInternalUsers": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
        },
        "gcimsCommitmentItemId": { "type": "string" },
        "formName": { "type": "string" },
        "termsConditionsUrlEN": { "type": "string" },
        "termsConditionsUrlFR": { "type": "string" },
        "contactEmail": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FundingOpportunityExpectedResult": {
      "type": "object",
      "properties": {
        "fundingOpportunityExpectedResultId": {
          "format": "uuid",
          "type": "string"
        },
        "expectedResultId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FundingOpportunityObjective": {
      "required": [ "fundingOpportunityId", "objectiveId" ],
      "type": "object",
      "properties": {
        "fundingOpportunityObjectiveId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "objectiveId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FundingOpportunityConsideration": {
      "type": "object",
      "properties": {
        "fundingOpportunityConsiderationId": {
          "format": "uuid",
          "type": "string"
        },
        "considerationId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FundingOpportunityFrequentlyAskedQuestion": {
      "type": "object",
      "properties": {
        "fundingOpportunityFrequentlyAskedQuestionId": {
          "format": "uuid",
          "type": "string"
        },
        "frequentlyAskedQuestionId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FundingOpportunityResource": {
      "type": "object",
      "properties": {
        "fundingOpportunityResourceId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "titleE": { "type": "string" },
        "titleF": { "type": "string" },
        "urlE": { "type": "string" },
        "urlF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}